simplify server.ts
