+++ backend/src/server.ts
@@ -1,105 +1,12 @@
 import express from 'express';
 import WebSocket from 'ws';
-
-import * as code from './code';
-import * as user from './user';
 import { handleConnection } from './connection';
 
 const app = express();
 app.use(express.json());
-const usedGame: string[] = [];
-// used ids for both players and host
-const usedId: string[] = [];
-// first key is gameId
-const games: Map<string, any> = new Map();
-let open = -1;
 
-app.get('/', (_req, res) => {
-	res.send('Hello, world!');
-});
-
-app.post('/games', (req, res) => {
-	try {
-		let quizData = JSON.parse(req.body);
-		const id = {
-			gameId: code.gen(5, usedGame),
-			hostId: code.gen(8, usedId),
-		};
-		// associate gameId with data and host
-		const data = { quiz: quizData, host: id.hostId };
-		games.set(id.gameId, data);
-		res.send(id);
-	} catch (e) {
-		// client error
-		console.error('Invalid JSON file:', e);
-		res.status(400).send('Invalid JSON file');
-	}
-});
-
-app.post('/games/:gameId/players', (_req, res) => {
-	user.userHandle(_req.params.gameId, _req, res);
-});
-
-app.get('/games/:id/questions/:index/start', (req, res) => {
-	const gameId = req.query.gameId as string;
-	const hostId = req.query.hostId as string;
-	const index = parseInt(req.params.index);
-	const game = games.get(gameId);
-
-	// client-requested game error
-	if (game === undefined) {
-		res.status(404).send(`Game ${gameId} not found`);
-		return;
-	}
-
-	const host = game.host;
-	// client permission error
-	if (host !== hostId) {
-		res.status(403).send(`Incorrect host of game ${gameId}`);
-		return;
-	}
-
-	const quiz = game.quiz;
-	// out-of-bounds error
-	if (index >= quiz.questions.length) {
-		res.status(404).send(`Question ${index} not found`);
-		return;
-	}
-
-	// start accepting answers for the question index
-	open = index;
-
-	// show question text and answers on both host and player screens? 
-
-	res.send({ ok: true });
-});
-
-app.post('/games/:id/questions/:index/answer', (req, res) => {
-	const gameId = req.query.gameId as string;
-	const playerId = req.query.playerId as string;
-	const index = parseInt(req.params.index);
-	const game = games.get(gameId);
-
-	// client-requested game error
-	if (game === undefined) {
-		res.status(404).send(`Game ${gameId} not found`);
-		return;
-	}
-
-	const quiz = game.quiz;
-	const question = quiz.questions[index];
-
-	// check if question is open
-	if (open != index) {
-		res.status(400).send(`Question ${index} is not open for answers`);
-		return;
-	}
-
-	// add map w/ player ids and answers?
-	const answer = req.body.answer;
-
-	// not accepting answers for the question index from the playerId 
-});
+import registerGameRoutes from './game';
+registerGameRoutes(app);
 
 // create websocket "server" which really piggybacks on the express server
 const webSocketServer = new WebSocket.Server({
@@ -111,7 +18,6 @@ const webSocketServer = new WebSocket.Server({
 	// certain connections from express to ws
 	noServer: true,
 });
-
 const httpServer = app.listen(8080);
 
 httpServer.on('listening', () => {
